php: helpers fixed
