+++ web/documentserver-example/php/doceditor.php
@@ -4,8 +4,10 @@ namespace OnlineEditorsExamplePhp;
 
 use OnlineEditorsExamplePhp\Helpers\ConfigManager;
 use OnlineEditorsExamplePhp\Helpers\ExampleUsers;
+use OnlineEditorsExamplePhp\Helpers\FileUtility;
 use OnlineEditorsExamplePhp\Helpers\JwtManager;
 use OnlineEditorsExamplePhp\Helpers\Users;
+use OnlineEditorsExamplePhp\Helpers\Utils;
 
 /**
  * (c) Copyright Ascensio System SIA 2023
@@ -29,13 +31,18 @@ require_once dirname(__FILE__) . '/vendor/autoload.php';
 
 $users = new ExampleUsers();
 $configManager = new ConfigManager();
+$fileUtility = new FileUtility();
+$utils = new Utils();
 $user = $users->getUser($_GET["user"]);
 $isEnableDirectUrl = isset($_GET["directUrl"]) ? filter_var($_GET["directUrl"], FILTER_VALIDATE_BOOLEAN) : false;
 
 // get the file url and upload it
 $externalUrl = $_GET["fileUrl"] ?? "";
 if (!empty($externalUrl)) {
-    $filename = doUpload($externalUrl);
+    try {
+        $filename = $utils->doUpload($externalUrl);
+    } catch (\Exception $e) {
+    }
 } else { // if the file url doesn't exist, get file name and file extension
     $filename = basename($_GET["fileID"]);
 }
@@ -43,7 +50,7 @@ $createExt = $_GET["fileExt"] ?? "";
 
 if (!empty($createExt)) {
     // and get demo file name by the extension
-    $filename = tryGetDefaultByType($createExt, $user);
+    $filename = $fileUtility->tryGetDefaultByType($createExt, $user);
 
     // create the demo file url
     $new_url = "doceditor.php?fileID=" . $filename . "&user=" . $_GET["user"];
@@ -53,16 +60,16 @@ if (!empty($createExt)) {
 
 $fileuri = fileUri($filename, true);
 $fileuriUser = realpath($configManager->getConfig("storagePath")) === $configManager->getConfig("storagePath") ?
-    getDownloadUrl($filename) . "&dmode=emb" : fileUri($filename);
-$directUrl = getDownloadUrl($filename, false);
-$docKey = getDocEditorKey($filename);
+    $fileUtility->getDownloadUrl($filename) . "&dmode=emb" : fileUri($filename);
+$directUrl = $fileUtility->getDownloadUrl($filename, false);
+$docKey = $fileUtility->getDocEditorKey($filename);
 $filetype = mb_strtolower(pathinfo($filename, PATHINFO_EXTENSION));
 
 $ext = mb_strtolower('.' . pathinfo($filename, PATHINFO_EXTENSION));
 $editorsMode = empty($_GET["action"]) ? "edit" : $_GET["action"];  // get the editors mode
 $canEdit = in_array($ext, $configManager->getConfig("docServEdited"));  // check if the file can be edited
 if ((!$canEdit && $editorsMode == "edit"
-    || $editorsMode == "fillForms")
+        || $editorsMode == "fillForms")
     && in_array($ext, $configManager->getConfig("docServFillforms"))
 ) {
     $editorsMode = "fillForms";
@@ -74,8 +81,8 @@ $submitForm = $editorsMode == "fillForms" && $user->id == "uid-1" && !1;
 $mode = $canEdit && $editorsMode != "view" ? "edit" : "view";  // define if the editing mode is edit or view
 $type = empty($_GET["type"]) ? "desktop" : $_GET["type"];
 
-$templatesImageUrl = getTemplateImageUrl($filename); // templates image url in the "From Template" section
-$createUrl = getCreateUrl($filename, $user->id, $type);
+$templatesImageUrl = $fileUtility->getTemplateImageUrl($filename); // templates image url in the "From Template" section
+$createUrl = $fileUtility->getCreateUrl($filename, $user->id, $type);
 $templates = [
     [
         "image" => "",
@@ -92,10 +99,10 @@ $templates = [
 // specify the document config
 $config = [
     "type" => $type,
-    "documentType" => getDocumentType($filename),
+    "documentType" => $fileUtility->getDocumentType($filename),
     "document" => [
         "title" => $filename,
-        "url" => getDownloadUrl($filename),
+        "url" => $fileUtility->getDownloadUrl($filename),
         "directUrl" => $isEnableDirectUrl ? $directUrl : "",
         "fileType" => $filetype,
         "key" => $docKey,
@@ -127,7 +134,7 @@ $config = [
         "actionLink" => empty($_GET["actionLink"]) ? null : json_decode($_GET["actionLink"]),
         "mode" => $mode,
         "lang" => empty($_COOKIE["ulang"]) ? "en" : $_COOKIE["ulang"],
-        "callbackUrl" => getCallbackUrl($filename),  // absolute URL to the document storage service
+        "callbackUrl" => $fileUtility->getCallbackUrl($filename),  // absolute URL to the document storage service
         "coEditing" => $editorsMode == "view" && $user->id == "uid-0" ? [
             "mode" => "strict",
             "change" => false,
@@ -158,7 +165,7 @@ $config = [
             "goback" => [  // settings for the Open file location menu button and upper right corner button
                 // the absolute URL to the website address which will be opened
                 // when clicking the Open file location menu button
-                "url" => serverPath(),
+                "url" => $fileUtility->serverPath(),
             ],
         ],
     ],
@@ -167,31 +174,31 @@ $config = [
 // an image for inserting
 $dataInsertImage = $isEnableDirectUrl ? [
     "fileType" => "png",
-    "url" => serverPath(true) . "/css/images/logo.png",
-    "directUrl" => serverPath(false) . "/css/images/logo.png",
+    "url" => $fileUtility->serverPath(true) . "/css/images/logo.png",
+    "directUrl" => $fileUtility->serverPath(false) . "/css/images/logo.png",
 ] : [
     "fileType" => "png",
-    "url" => serverPath(true) . "/css/images/logo.png",
+    "url" => $fileUtility->serverPath(true) . "/css/images/logo.png",
 ];
 
 // a document for comparing
 $dataCompareFile = $isEnableDirectUrl ? [
     "fileType" => "docx",
-    "url" => serverPath(true) . "/webeditor-ajax.php?type=assets&name=sample.docx",
-    "directUrl" => serverPath(false) . "/webeditor-ajax.php?type=assets&name=sample.docx",
+    "url" => $fileUtility->serverPath(true) . "/webeditor-ajax.php?type=assets&name=sample.docx",
+    "directUrl" => $fileUtility->serverPath(false) . "/webeditor-ajax.php?type=assets&name=sample.docx",
 ] : [
     "fileType" => "docx",
-    "url" => serverPath(true) . "/webeditor-ajax.php?type=assets&name=sample.docx",
+    "url" => $fileUtility->serverPath(true) . "/webeditor-ajax.php?type=assets&name=sample.docx",
 ];
 
 // recipients data for mail merging
 $dataMailMergeRecipients = $isEnableDirectUrl ? [
     "fileType" => "csv",
-    "url" => serverPath(true) . "/webeditor-ajax.php?type=csv",
-    "directUrl" => serverPath(false) . "/webeditor-ajax.php?type=csv",
+    "url" => $fileUtility->serverPath(true) . "/webeditor-ajax.php?type=csv",
+    "directUrl" => $fileUtility->serverPath(false) . "/webeditor-ajax.php?type=csv",
 ] : [
     "fileType" => "csv",
-    "url" => serverPath(true) . "/webeditor-ajax.php?type=csv",
+    "url" => $fileUtility->serverPath(true) . "/webeditor-ajax.php?type=csv",
 ];
 
 // users data for mentions
@@ -200,234 +207,13 @@ $usersForMentions = $user->id != "uid-0" ? $users->getUsersForMentions($user->id
 // check if the secret key to generate token exists
 $jwtManager = new JwtManager();
 if ($jwtManager->isJwtEnabled()) {
-    $config["token"] = jwtEncode($config);  // encode config into the token
+    $config["token"] = $jwtManager->jwtEncode($config);  // encode config into the token
     // encode the dataInsertImage object into the token
     $dataInsertImage["token"] = $jwtManager->jwtEncode($dataInsertImage);
     // encode the dataCompareFile object into the token
     $dataCompareFile["token"] = $jwtManager->jwtEncode($dataCompareFile);
     // encode the dataMailMergeRecipients object into the token
     $dataMailMergeRecipients["token"] = $jwtManager->jwtEncode($dataMailMergeRecipients);
-}
-
-/**
- * Get demo file name by the extension
- *
- * @param string $createExt
- * @param Users $user
- *
- * @return string
- */
-function tryGetDefaultByType($createExt, $user)
-{
-    $demoName = ($_GET["sample"] ? "sample." : "new.") . $createExt;
-    $demoPath = "assets" . DIRECTORY_SEPARATOR . ($_GET["sample"] ? "sample" : "new") . DIRECTORY_SEPARATOR;
-    $demoFilename = GetCorrectName($demoName);
-
-    if (!@copy(dirname(__FILE__) . DIRECTORY_SEPARATOR . $demoPath . $demoName, getStoragePath($demoFilename))) {
-        sendlog("Copy file error to ". getStoragePath($demoFilename), "common.log");
-        // Copy error!!!
-    }
-
-    // create demo file meta information
-    createMeta($demoFilename, $user->id, $user->name);
-
-    return $demoFilename;
-}
-
-/**
- * Get the callback url
- *
- * @param string $fileName
- *
- * @return string
- */
-function getCallbackUrl($fileName)
-{
-    return serverPath(true) . '/'
-                . "webeditor-ajax.php"
-                . "?type=track"
-                . "&fileName=" . urlencode($fileName)
-                . "&userAddress=" . getClientIp();
-}
-
-/**
- * Get url to the created file
- *
- * @param string $fileName
- * @param string $uid
- * @param string $type
- *
- * @return string
- */
-function getCreateUrl($fileName, $uid, $type)
-{
-    $ext = trim(getInternalExtension($fileName), '.');
-    return serverPath(false) . '/'
-            . "doceditor.php"
-            . "?fileExt=" . $ext
-            . "&user=" . $uid
-            . "&type=" . $type;
-}
-
-/**
- * Get url for history download
- *
- * @param string $fileName
- * @param string $version
- * @param string $file
- * @param bool $isServer
- *
- * @return string
- */
-function getHistoryDownloadUrl($fileName, $version, $file, $isServer = true)
-{
-    $userAddress = $isServer ? "&userAddress=" . getClientIp() : "";
-    return serverPath($isServer) . '/'
-        . "webeditor-ajax.php"
-        . "?type=history"
-        . "&fileName=" . urlencode($fileName)
-        . "&ver=" . $version
-        . "&file=" . urlencode($file)
-        . $userAddress;
-}
-
-/**
- * Get url to download a file
- *
- * @param string $fileName
- * @param bool $isServer
- *
- * @return string
- */
-function getDownloadUrl($fileName, $isServer = true)
-{
-    $userAddress = $isServer ? "&userAddress=" . getClientIp() : "";
-    return serverPath($isServer) . '/'
-        . "webeditor-ajax.php"
-        . "?type=download"
-        . "&fileName=" . urlencode($fileName)
-        . $userAddress;
-}
-
-/**
- * Get document history
- *
- * @param string $filename
- * @param string $filetype
- * @param string $docKey
- * @param string $fileuri
- * @param bool $isEnableDirectUrl
- *
- * @return array
- */
-function getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl)
-{
-    $configManager = new ConfigManager();
-    $storagePath = $configManager->getConfig("storagePath");
-    $histDir = getHistoryDir(getStoragePath($filename));  // get the path to the file history
-
-    if (getFileVersion($histDir) > 0) {  // check if the file was modified (the file version is greater than 0)
-        $curVer = getFileVersion($histDir);
-
-        $hist = [];
-        $histData = [];
-
-        for ($i = 1; $i <= $curVer; $i++) {  // run through all the file versions
-            $obj = [];
-            $dataObj = [];
-            $verDir = getVersionDir($histDir, $i);  // get the path to the file version
-
-            // get document key
-            $key = $i == $curVer ? $docKey : file_get_contents($verDir . DIRECTORY_SEPARATOR . "key.txt");
-            $obj["key"] = $key;
-            $obj["version"] = $i;
-
-            if ($i == 1) {  // check if the version number is equal to 1
-                // get meta data of this file
-                $createdInfo = file_get_contents($histDir . DIRECTORY_SEPARATOR . "createdInfo.json");
-                $json = json_decode($createdInfo, true);  // decode the meta data from the createdInfo.json file
-
-                $obj["created"] = $json["created"];
-                $obj["user"] = [
-                    "id" => $json["uid"],
-                    "name" => $json["name"],
-                ];
-            }
-
-            $fileExe = mb_strtolower(pathinfo($filename, PATHINFO_EXTENSION));
-
-            $prevFileName = $verDir . DIRECTORY_SEPARATOR . "prev." . $filetype;
-            $prevFileName = mb_substr($prevFileName, mb_strlen(getStoragePath("")));
-            $dataObj["fileType"] = $fileExe;
-            $dataObj["key"] = $key;
-
-            $directUrl = $i == $curVer ? fileUri($filename, false) :
-                getHistoryDownloadUrl($filename, $i, "prev.".$fileExe, false);
-            $prevFileUrl = $i == $curVer ? $fileuri : getHistoryDownloadUrl($filename, $i, "prev.".$fileExe);
-            if (realpath($storagePath) === $storagePath) {
-                $prevFileUrl = $i == $curVer ? getDownloadUrl($filename) :
-                    getHistoryDownloadUrl($filename, $i, "prev.".$fileExe);
-                if ($isEnableDirectUrl) {
-                    $directUrl = $i == $curVer ? getDownloadUrl($filename, false) :
-                        getHistoryDownloadUrl($filename, $i, "prev.".$fileExe, false);
-                }
-            }
-
-            $dataObj["url"] = $prevFileUrl;  // write file url to the data object
-            if ($isEnableDirectUrl) {
-                $dataObj["directUrl"] = $directUrl;  // write direct url to the data object
-            }
-            $dataObj["version"] = $i;
-
-            if ($i > 1) {  // check if the version number is greater than 1 (the document was modified)
-                $changes = json_decode(file_get_contents(getVersionDir($histDir, $i - 1) .
-                    DIRECTORY_SEPARATOR . "changes.json"), true);  // get the path to the changes.json file
-                $change = $changes["changes"][0];
-
-                // write information about changes to the object
-                $obj["changes"] = $changes ? $changes["changes"] : null;
-                $obj["serverVersion"] = $changes["serverVersion"];
-                $obj["created"] = $change ? $change["created"] : null;
-                $obj["user"] = $change ? $change["user"] : null;
-
-                $prev = $histData[$i - 2];  // get the history data from the previous file version
-                // write information about previous file version to the data object
-                $dataObj["previous"] = $isEnableDirectUrl ? [
-                    "fileType" => $prev["fileType"],
-                    "key" => $prev["key"],
-                    "url" => $prev["url"],
-                    "directUrl" => $prev["directUrl"],
-                ] : [
-                    "fileType" => $prev["fileType"],
-                    "key" => $prev["key"],
-                    "url" => $prev["url"],
-                ];
-
-                // write the path to the diff.zip archive with differences in this file version
-                $dataObj["changesUrl"] = getHistoryDownloadUrl($filename, $i - 1, "diff.zip");
-            }
-
-            $jwtManager = new JwtManager();
-            if ($jwtManager->isJwtEnabled()) {
-                $dataObj["token"] = $jwtManager->jwtEncode($dataObj);
-            }
-
-            $hist[] = $obj;  // add object dictionary to the hist list
-            $histData[$i - 1] = $dataObj;  // write data object information to the history data
-        }
-
-        // write history information about the current file version
-        $out = [];
-        array_push(
-            $out,
-            [
-                "currentVersion" => $curVer,
-                "history" => $hist,
-            ],
-            $histData
-        );
-        return $out;
-    }
 }
 
 ?>
@@ -441,7 +227,7 @@ function getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl)
             maximum-scale=1, minimum-scale=1, user-scalable=no, minimal-ui" />
     <meta name="apple-mobile-web-app-capable" content="yes" />
     <meta name="mobile-web-app-capable" content="yes" />
-    <link rel="icon" href="css/images/<?php echo getDocumentType($filename) ?>.ico" type="image/x-icon" />
+    <link rel="icon" href="css/images/<?php echo $fileUtility->getDocumentType($filename) ?>.ico" type="image/x-icon" />
     <title>ONLYOFFICE</title>
 
     <style>
@@ -475,7 +261,7 @@ function getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl)
 
     <script type="text/javascript" src="
         <?php echo $configManager->getConfig("docServSiteUrl").
-            $configManager->getConfig("docServApiUrl") ?>">
+        $configManager->getConfig("docServApiUrl") ?>">
     </script>
 
     <script type="text/javascript">
@@ -617,11 +403,11 @@ function getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl)
 
         var сonnectEditor = function () {
 
-        <?php
-        if (!file_exists(getStoragePath($filename))) {
-            echo "alert('File not found'); return;";
-        }
-        ?>
+            <?php
+            if (!file_exists($fileUtility->getStoragePath($filename))) {
+                echo "alert('File not found'); return;";
+            }
+            ?>
 
             config = <?php echo json_encode($config) ?>;
 
@@ -642,44 +428,44 @@ function getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl)
             };
 
             <?php
-            $out = getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl);
+            $out = $fileUtility->getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl);
             $history = $out[0];
             $historyData = $out[1];
             ?>
 
             <?php if ($user->id != "uid-0") { ?>
-                <?php if ($history != null && $historyData != null) { ?>
-                    // the user is trying to show the document version history
-                    config.events['onRequestHistory'] = function () {
-                        // show the document version history
-                        docEditor.refreshHistory(<?php echo json_encode($history) ?>);
-                    };
-                    // the user is trying to click the specific document version in the document version history
-                    config.events['onRequestHistoryData'] = function (event) {
-                        var ver = event.data;
-                        var histData = <?php echo json_encode($historyData) ?>;
-                        // send the link to the document for viewing the version history
-                        docEditor.setHistoryData(histData[ver - 1]);
-                    };
-                    // the user is trying to go back to the document from viewing the document version history
-                    config.events['onRequestHistoryClose'] = function () {
-                        document.location.reload();
-                    };
-                <?php } ?>
-                // add mentions for not anonymous users
-                config.events['onRequestUsers'] = function () {
-                    docEditor.setUsers({  // set a list of users to mention in the comments
-                        "users": <?php echo json_encode($usersForMentions) ?>
-                    });
-                };
-                // the user is mentioned in a comment
-                config.events['onRequestSendNotify'] = function (event) {
-                    event.data.actionLink = replaceActionLink(location.href, JSON.stringify(event.data.actionLink));
-                    var data = JSON.stringify(event.data);
-                    innerAlert("onRequestSendNotify: " + data);
-                };
-                // prevent file renaming for anonymous users
-                config.events['onRequestRename'] = onRequestRename;
+            <?php if ($history != null && $historyData != null) { ?>
+            // the user is trying to show the document version history
+            config.events['onRequestHistory'] = function () {
+                // show the document version history
+                docEditor.refreshHistory(<?php echo json_encode($history) ?>);
+            };
+            // the user is trying to click the specific document version in the document version history
+            config.events['onRequestHistoryData'] = function (event) {
+                var ver = event.data;
+                var histData = <?php echo json_encode($historyData) ?>;
+                // send the link to the document for viewing the version history
+                docEditor.setHistoryData(histData[ver - 1]);
+            };
+            // the user is trying to go back to the document from viewing the document version history
+            config.events['onRequestHistoryClose'] = function () {
+                document.location.reload();
+            };
+            <?php } ?>
+            // add mentions for not anonymous users
+            config.events['onRequestUsers'] = function () {
+                docEditor.setUsers({  // set a list of users to mention in the comments
+                    "users": <?php echo json_encode($usersForMentions) ?>
+                });
+            };
+            // the user is mentioned in a comment
+            config.events['onRequestSendNotify'] = function (event) {
+                event.data.actionLink = replaceActionLink(location.href, JSON.stringify(event.data.actionLink));
+                var data = JSON.stringify(event.data);
+                innerAlert("onRequestSendNotify: " + data);
+            };
+            // prevent file renaming for anonymous users
+            config.events['onRequestRename'] = onRequestRename;
             <?php } ?>
 
             if (config.editorConfig.createUrl) {
@@ -704,9 +490,9 @@ function getHistory($filename, $filetype, $docKey, $fileuri, $isEnableDirectUrl)
     </script>
 </head>
 <body>
-    <form id="form1">
-        <div id="iframeEditor">
-        </div>
-    </form>
+<form id="form1">
+    <div id="iframeEditor">
+    </div>
+</form>
 </body>
 </html>
